sail_netstack

Module ip

Structs

FragmentExpirations

FragmentReassembler

FragmentStats

ParsedIcmpPacket

PmtuStats

PmtuTable

Enums

FragmentError

IcmpErrorKind

IcmpMessage

PmtuError

Functions

emit_icmp_echo_reply

Emits an echo reply with the request payload and reversed IP addresses.

emit_icmp_error

Emits an ICMP error quoting as much of the invoking packet as is safe for the minimum reassembly size. Errors are never generated in response to multicast/broadcast destinations, non-initial fragments, or ICMP errors.

fragment_outbound_ip_packet

Fragments a locally generated IPv4 or IPv6 packet to mtu.

parse_icmp_packet

Parses an ICMP message after IP reassembly.